Attempt at fixing a thread-safety issue in ManagedSpriteInstance which was causing, somehow, an invalid list reference to be used. Removing TRACE logs from Shipping builds. Adding more exception types and ThrowHelpers. Fixing outlining issue with overridden stroke text. Making further changes to premultiplied alpha when fully-translucent objects are used. Adding combined xBRZ + EPX resampler. Changing EPX to use a proper color comparer, adding EPXLegacy with old comparer. Adding common logic for scalers to remove a lot of boilerplate code.

+++ SpriteMaster/Configuration/Config.cs
@@ -602,7 +602,7 @@ internal static class Config {
 			[Attributes.Comment("The tolerance for colors to be considered equal - [0, 256)")]
 			[Attributes.OptionsAttribute(Attributes.OptionsAttribute.Flag.FlushAllInternalCaches)]
 			[Attributes.LimitsInt(0, 255)]
-			internal static int EqualColorTolerance = 20;
+			internal static byte EqualColorTolerance = 20;
 			[Attributes.Comment("The weight provided to luminance as opposed to chrominance when performing color comparisons")]
 			[Attributes.OptionsAttribute(Attributes.OptionsAttribute.Flag.FlushAllInternalCaches)]
 			[Attributes.LimitsReal(0.0, 10.0)]
